+++ lib/input.rb
@@ -3,7 +3,7 @@ $clipboard = nil
 
 class Input
   attr_sprite
-  attr_reader :value, :selection_start, :selection_end
+  attr_reader :value, :selection_start, :selection_end, :lines
 
   SIZE_ENUM = {
     small: -1,
@@ -17,7 +17,30 @@ class Input
   CURSOR_FULL_TICKS = 30
   CURSOR_FLASH_TICKS = 20
 
+  NOOP = -> {}
+
+  META_KEYS = %i[meta_left meta_right] # and `meta`
+  SHIFT_KEYS = %i[shift_left shift_right]
+  ALT_KEYS = %i[alt_left alt_right]
+  CTRL_KEYS = %i[control_left control_right]
+  DEL_KEYS = %i[delete backspace]
+  IGNORE_KEYS = %i[raw_key char meta shift alt control] + META_KEYS + SHIFT_KEYS + ALT_KEYS + CTRL_KEYS
+
   @@id = 0
+
+  class Line
+    attr_reader :number, :text, :start
+
+    def initialize(number, start, text)
+      @number = number
+      @start = start
+      @text = text
+    end
+
+    def length
+      @text.length
+    end
+  end
 
   def initialize(**params)
     @x = params[:x] || 0
@@ -29,7 +52,7 @@ class Input
     @word_chars = params[:word_chars] || ('a'..'z').to_a + ('A'..'Z').to_a + ('0'..'9').to_a + ['_', '-']
     _, @font_height = $gtk.calcstringbox(@word_chars.join(''), @size_enum, @font)
     @punctuation_chars = params[:punctuation_chars] || %w[! % , . ; : ' " ` ) \] } * &]
-    @crnl_chars = ["\r", "\n"]
+    @crlf_chars = ["\r", "\n"]
     @word_wrap_chars = @word_chars + @punctuation_chars
 
     @padding = params[:padding] || 2
@@ -62,7 +85,6 @@ class Input
     @cursor_ticks = 0
     @cursor_dir = 1
 
-    # TODO: implement key repeat for cursor movement
     @key_repeat_delay = params[:key_repeat_delay] = 20
     @key_repeat_debounce = params[:key_repeat_debounce] = 5
 
@@ -72,8 +94,14 @@ class Input
     # Render target for text scrolling
     @path = "__input_#{@@id += 1}"
     @source_x = 0
+    @source_y = 0
 
     @word_wrap = params[:word_wrap] || false
+    @focussed = params[:focussed] || false
+    @will_focus = false # Get the focus at the end of the tick
+
+    @on_clicked = params[:on_clicked] || NOOP
+    @on_unhandled_key = params[:on_unhandled_key] || NOOP
   end
 
   def draw_override(ffi)
@@ -107,38 +135,58 @@ class Input
 
     # CURSOR
     # TODO: Cursor renders outside of the bounds of the control
-    @cursor_ticks += @cursor_dir
-    alpha = if @cursor_ticks == CURSOR_FULL_TICKS
-              @cursor_dir = -1
-              255
-            elsif @cursor_ticks == 0
-              @cursor_dir = 1
-              0
-            elsif @cursor_ticks < CURSOR_FULL_TICKS
-              $args.easing.ease(0, @cursor_ticks, CURSOR_FLASH_TICKS, :quad) * 255
-            else
-              255
-            end
-    ffi.draw_solid(@cursor_x, @cursor_y, @padding, @font_height + @padding * 2, 0, 0, 0, alpha)
+    if @focussed
+      @cursor_ticks += @cursor_dir
+      alpha = if @cursor_ticks == CURSOR_FULL_TICKS
+                @cursor_dir = -1
+                255
+              elsif @cursor_ticks == 0
+                @cursor_dir = 1
+                0
+              elsif @cursor_ticks < CURSOR_FULL_TICKS
+                $args.easing.ease(0, @cursor_ticks, CURSOR_FLASH_TICKS, :quad) * 255
+              else
+                255
+              end
+      ffi.draw_solid(@cursor_x, @cursor_y, @padding, @font_height + @padding * 2, 0, 0, 0, alpha)
+    end
+
+    if @will_focus
+      @will_focus = false
+      @focussed = true
+    end
   end
 
-  META_KEYS = %i[meta_left meta_right] # and `meta`
-  SHIFT_KEYS = %i[shift_left shift_right]
-  ALT_KEYS = %i[alt_left alt_right]
-  CTRL_KEYS = %i[control_left control_right]
-  DEL_KEYS = %i[delete backspace]
-
   def tick
-    prepare_special_keys
-    handle_keyboard
+    if @focussed
+      prepare_special_keys
+      handle_keyboard
+    end
     handle_mouse
     prepare_render_target
+  end
+
+  def focussed?
+    @focussed
+  end
+
+  def focus!
+    @will_focus = true
+  end
+
+  def blur!
+    @focussed = false
   end
 
   def prepare_special_keys
     keyboard = $args.inputs.keyboard
 
-    @down_keys = keyboard.key_down.truthy_keys
+    tick_count = $args.tick_count
+    repeat_keys = keyboard.key_held.truthy_keys.select do |key|
+      ticks = tick_count - keyboard.key_held.send(key).to_i
+      ticks > @key_repeat_delay && ticks % @key_repeat_debounce == 0
+    end
+    @down_keys = keyboard.key_down.truthy_keys.concat(repeat_keys) - IGNORE_KEYS
 
     # Find special keys
     special_keys = keyboard.key_down.truthy_keys + keyboard.key_held.truthy_keys
@@ -156,17 +204,13 @@ class Input
       if @down_keys.include?(:a)
         @selection_start = 0
         @selection_end = @value.length
-      end
-
-      if @down_keys.include?(:c) && @selection_start != @selection_end
+      elsif @down_keys.include?(:c) && @selection_start != @selection_end
         $clipboard = if @selection_start < @selection_end
                        @value.slice(@selection_start, @selection_end - @selection_start)
                      else
                        @value.slice(@selection_end, @selection_start - @selection_end)
                      end
-      end
-
-      if @down_keys.include?(:x) && @selection_start != @selection_end
+      elsif @down_keys.include?(:x) && @selection_start != @selection_end
         $clipboard = if @selection_start < @selection_end
                        @value.slice(@selection_start, @selection_end - @selection_start)
                      else
@@ -174,27 +218,35 @@ class Input
                      end
         @value = @value.slice(0, @selection_start.lesser(@selection_end)) + @value.slice(@selection_end.greater(@selection_start), @value.length)
         @selection_start = @selection_end = @selection_start.lesser(@selection_end)
-      end
-
-      if @down_keys.include?(:v)
+      elsif @down_keys.include?(:v)
         @value = @value.slice(0, @selection_start.lesser(@selection_end)) + $clipboard + @value.slice(@selection_end.greater(@selection_start), @value.length)
         @selection_start = @selection_end = @selection_start.lesser(@selection_end) + $clipboard.length
-      end
-
-      if @down_keys.include?(:left)
+      elsif @down_keys.include?(:left)
+        index = if @word_wrap
+                  find_line.start
+                else
+                  0
+                end
         if @shift
-          @selection_end = 0
+          @selection_end = index
         else
-          @selection_start = @selection_end = 0
+          @selection_start = @selection_end = index
         end
-      end
-
-      if @down_keys.include?(:right)
+      elsif @down_keys.include?(:right)
+        index = if @word_wrap
+                  line = find_line
+                  putz line
+                  line.start + line.length
+                else
+                  @value.length
+                end
         if @shift
-          @selection_end = @value.length
+          @selection_end = index
         else
-          @selection_start = @selection_end = @value.length
+          @selection_start = @selection_end = index
         end
+      else
+        @on_unhandled_key.call(@down_keys.first, self)
       end
     elsif text_keys.empty?
       if (@down_keys & DEL_KEYS).any?
@@ -247,8 +299,54 @@ class Input
                              end
           @selection_end = @selection_start
         end
+      elsif @down_keys.include?(:up) && @word_wrap
+        if @shift
+          line = find_line
+          @selection_end = if line.number == 0
+                             0
+                           else
+                             prev = @lines[line.number - 1]
+                             line.start - prev.length + find_index_at_x(@cursor_x - @x + @source_x, prev) - 1
+                           end
+        else
+          @selection_start = if @alt
+                               # TODO: beginning of previous paragraph
+                             else
+                               line = find_line
+                               if line.number == 0
+                                 0
+                               else
+                                 prev = @lines[line.number - 1]
+                                 line.start - prev.length + find_index_at_x(@cursor_x - @x + @source_x, prev) - 1
+                               end
+                             end
+          @selection_end = @selection_start
+        end
+      elsif @down_keys.include?(:down) && @word_wrap
+        if @shift
+          line = find_line
+          @selection_end = if line.number == @lines.length - 1
+                             @value.length
+                           else
+                             find_index_at_x(@cursor_x - @x + @source_x, @lines[line.number + 1]) + line.start + line.length - 1
+                           end
+        else
+          @selection_start = if @alt
+                               # TODO: end of next paragraph
+                             else
+                               line = find_line
+                               if line.number == @lines.length - 1
+                                 @value.length
+                               else
+                                 find_index_at_x(@cursor_x - @x + @source_x, @lines[line.number + 1]) + line.start + line.length - 1
+                               end
+                             end
+          @selection_end = @selection_start
+        end
       elsif @down_keys.include?(:enter) && @word_wrap
         insert("\n")
+      else
+        @on_unhandled_key.call(@down_keys.first, self)
       end
     else
       insert(text_keys.join(''))
@@ -274,34 +372,47 @@ class Input
     mouse = $args.inputs.mouse
 
     if !@mouse_down && mouse.down && mouse.inside_rect?(self)
+      @on_clicked.call(mouse, self)
+      return unless @focussed || @will_focus
+
       @mouse_down = true
 
-      index = find_index_at_x(mouse.x - @x + @source_x)
+      index = if @word_wrap
+                line = (@h + @y - mouse.y + @source_y).idiv(@font_height).cap_min_max(0, @lines.length - 1)
+                find_index_at_x(mouse.x - @x + @source_x, @lines[line]) + lines[0, line].sum(&:length)
+              else
+                find_index_at_x(mouse.x - @x + @source_x)
+              end
       if @shift
         @selection_end = index
       else
         @selection_start = @selection_end = index
       end
     elsif @mouse_down
-      @selection_end = find_index_at_x(mouse.x - @x + @source_x)
+      index = if @word_wrap
+                line = (@h + @y - mouse.y + @source_y).idiv(@font_height).clamp(0, @lines.length - 1)
+                find_index_at_x(mouse.x - @x + @source_x, @lines[line]) + lines[0, line].sum(&:length)
+              else
+                find_index_at_x(mouse.x - @x + @source_x)
+              end
+      @selection_end = index
       @mouse_down = false if mouse.up
     end
   end
 
-  # BUG: Single character words are busted
   # TODO: Improve walking words
   def find_word_break_left
     return 0 if @selection_end == 0
 
     index = @selection_end
-    found_word_char = false
-    while !found_word_char
+
+    loop do
       index -= 1
       return 0 if index == 0
-      found_word_char = true if @word_chars.include?(@value[index, 1])
+      break if @word_chars.include?(@value[index, 1])
     end
 
-    while true
+    loop do
       index -= 1
       return 0 if index == 0
       return index + 1 unless @word_chars.include?(@value[index, 1])
@@ -312,14 +423,13 @@ class Input
     length = @value.length
     return length if index == length
 
-    found_word_char = false
-    while !found_word_char
+    loop do
       index += 1
       return length if index == length
-      found_word_char = true if @word_chars.include?(@value[index, 1])
+      break if @word_chars.include?(@value[index, 1])
     end
 
-    while true
+    loop do
       index += 1
       return length if index == length
       return index unless @word_chars.include?(@value[index, 1])
@@ -330,7 +440,7 @@ class Input
     # @word_chars = params[:word_chars] || ('a'..'z').to_a + ('A'..'Z').to_a + ('0'..'9').to_a + ['_', '-']
     # _, @font_height = $gtk.calcstringbox(@word_chars.join(''), @size_enum, @font)
     # @punctuation_chars = params[:punctuation_chars] || %w[! % , . ; : ' " ` ) \] } * &]
-    # @crnl_chars = ["\r", "\n"]
+    # @crlf_chars = ["\r", "\n"]
     # @word_wrap_chars = @word_chars + @punctuation_chars
     words = []
     word = ''
@@ -342,7 +452,7 @@ class Input
       case mode
       when :leading_white_space
         if value[index].strip == '' # leading white-space
-          if @crnl_chars.include?(value[index]) # TODO: prolly need to replace \r\n with \n up front
+          if @crlf_chars.include?(value[index]) # TODO: prolly need to replace \r\n with \n up front
             words << word
             word = "\n"
           else
@@ -355,7 +465,7 @@ class Input
       when :word_wrap_chars # TODO: consider smarter handling. "something!)something" would be considered a word right now, theres an extra step needed
         if @word_wrap_chars.include?(value[index])
           word << value[index]
-        elsif @crnl_chars.include?(value[index])
+        elsif @crlf_chars.include?(value[index])
           words << word
           word = "\n"
           mode = :leading_white_space
@@ -365,7 +475,7 @@ class Input
         end
       when :trailing_white_space
         if value[index].strip == '' # trailing white-space
-          if @crnl_chars.include?(value[index])
+          if @crlf_chars.include?(value[index])
             words << word
             word = "\n" # converting all new line chars to \n
             mode = :leading_white_space
@@ -388,7 +498,6 @@ class Input
     line = ''
     i = -1
     l = words.length
-    # ["1", "\n", "\n", "\n2"]
     while (i += 1) < l
       word = words[i]
       if word == "\n"
@@ -409,17 +518,24 @@ class Input
     end
 
     lines << line
-    # lines << '' if line.end_with?("\n")
-    # lines
   end
 
-  def find_index_at_x(x)
+  def find_line(index = @selection_end)
+    i = -1
+    start_index = 0
+    while @lines[i += 1].length + start_index < index
+      start_index += @lines[i].length
+    end
+    Line.new(i, start_index, @lines[i])
+  end
+
+  def find_index_at_x(x, str = @value)
     return 0 if x < @padding
 
     index = 0
-    while index < @value.length
+    while index < str.length
       index += 1
-      width, = $gtk.calcstringbox(@value[0, index].to_s, @size_enum, @font)
+      width, = $gtk.calcstringbox(str[0, index].to_s, @size_enum, @font)
       break if width > x
     end
     index
@@ -428,9 +544,9 @@ class Input
   def prepare_render_target
     if @word_wrap
       # calculate lines
-      lines = perform_word_wrap
+      @lines = perform_word_wrap
 
-      @h = lines.length * @font_height + 2 * @padding # TODO: Implement line spacing
+      @h = @lines.length * @font_height + 2 * @padding # TODO: Implement line spacing
       rt = $args.outputs[@path]
       rt.w = @w
       rt.h = @h
@@ -438,8 +554,8 @@ class Input
       # TODO: implement sprite background
       rt.transient!
 
-      putz @value.gsub("\n", '\n')
-      putz lines
+      # putz @value.gsub("\n", '\n')
+      # putz lines
 
       if @selection_start != @selection_end
         selection_start_count = @selection_start.lesser(@selection_end)
@@ -450,11 +566,12 @@ class Input
       end
       cursor_count = @selection_end
 
-      lines.each_with_index do |line, i|
+      @lines.each_with_index do |line, i|
         y = @h - @padding - (i + 1) * @font_height
-        puts "#{i}: #{line.gsub("\n", '\n')}"
 
         # SELECTION
+        # TODO: Show cursor at start of next line if after the white space at the end of the current line (0)
+        # TODO: Ensure cursor_x doesn't go past the line width
         if selection_start_count >= 0
           if selection_start_count - line.length <= 0
             # selection starts here